Document some internal `fxa-*` claims inside JWT access tokens.
Inherited from BrowserID, some internal Mozilla reliers need access
to special `fxa-*` prefixed claims. These were only documented in
the fxa-auth-server API docs about BrowserID, so I've copied the
text over to the OAuth docs to make it easier to find, and added
a few notes on how we expect the values to be used.