Fix numbering in OAuth dance explanation
This is rendering as an ordered list starting from 0 and the nested ordered list isn't rendering at all. I can't find anything about docusaurus that would explain this, and the GitHub preview is of course using its own renderer.

This patch switches to a (hopefully) more canonical method of indicating both list beginnings, which ideally will result in the lists being rendered correctly.